# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pharmacy Management System API",
        "description": "Complete API documentation for Pharmacy Management System including admin, student, teacher, and payment modules",
        "contact": {
            "email": "support@pharmacy.edu"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pharmacy.edu",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/admin/save-student-marks": {
            "post": {
                "tags": [
                    "Admin - Student Marks"
                ],
                "summary": "Save student marks (bulk)",
                "description": "Save marks for one or multiple students using fn_admin_savestudentmarks_v3 stored procedure",
                "operationId": "c17c1524c8b661b87af7679fb065e110",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "marks"
                                ],
                                "properties": {
                                    "marks": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "p_marks_id",
                                                "p_exam_status_code",
                                                "p_submit_type_id",
                                                "p_evaluator_type_id",
                                                "p_admin_user_id"
                                            ],
                                            "properties": {
                                                "p_marks_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "p_external_marks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 75,
                                                    "nullable": true
                                                },
                                                "p_internal_marks": {
                                                    "type": "number",
                                                    "format": "double",
                                                    "example": 28,
                                                    "nullable": true
                                                },
                                                "p_doc": {
                                                    "type": "string",
                                                    "example": "marksheet.pdf",
                                                    "nullable": true
                                                },
                                                "p_exam_status_code": {
                                                    "type": "string",
                                                    "example": "PASS"
                                                },
                                                "p_submit_type_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "p_evaluator_type_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "p_admin_user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "p_remarks": {
                                                    "type": "string",
                                                    "example": "Marks updated successfully",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marks saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "version": {
                                            "type": "string",
                                            "example": "1.0"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All 1 marks saved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "success": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "failed": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "results": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "marks_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "status": {
                                                                "type": "string",
                                                                "example": "success"
                                                            },
                                                            "result_code": {
                                                                "type": "integer",
                                                                "example": 0
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "Marks saved successfully"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized - Token missing or invalid"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Enter your authentication token",
                "name": "token",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}